+++ contracts/dao/AbstractController.sol
@@ -1,13 +1,11 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.10;
 
-
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/utils/math/Math.sol";
 import "@openzeppelin/contracts/utils/math/SafeMath.sol";
-import '@openzeppelin/contracts/utils/structs/EnumerableSet.sol';
+import "@openzeppelin/contracts/utils/structs/EnumerableSet.sol";
 import "@openzeppelin/contracts/access/Ownable.sol";
-
 
 import "../interface/IDistribute.sol";
 
@@ -15,48 +13,50 @@ import "../tools/CheckPermission.sol";
 import "../interface/IVeToken.sol";
 
 abstract contract AbstractController is CheckPermission {
-
     using SafeMath for uint256;
     using EnumerableSet for EnumerableSet.AddressSet;
 
-    event Attach(address indexed owner, address indexed gauge, uint tokenId);
-    event Detach(address indexed owner, address indexed gauge, uint tokenId);
-    event Voted(address indexed voter, uint tokenId, uint weight);
-    event Abstained(uint tokenId, uint weight);
+    event Attach(address indexed owner, address indexed gauge, uint256 tokenId);
+    event Detach(address indexed owner, address indexed gauge, uint256 tokenId);
+    event Voted(address indexed voter, uint256 tokenId, uint256 weight);
+    event Abstained(uint256 tokenId, uint256 weight);
 
     struct PoolVote {
         address pool;
-        uint lastUse;
+        uint256 lastUse;
     }
 
     address public immutable veToken; // the ve token that governs these contracts
     address public immutable base;
     address public immutable distribute;
 
-    uint public duration;
-    uint public totalWeight; // total voting weight
-    uint public lastUpdate;
+    uint256 public duration;
+    uint256 public totalWeight; // total voting weight
+    uint256 public lastUpdate;
 
-    mapping(address => uint) public weights; // pool => weight
-    mapping(uint => uint) public usedWeights;  // nft => total voting weight of user
-    mapping(uint => PoolVote) public userPool;  // nft => pool voting weight of user
+    mapping(address => uint256) public weights; // pool => weight
+    mapping(uint256 => uint256) public usedWeights; // nft => total voting weight of user
+    mapping(uint256 => PoolVote) public userPool; // nft => pool voting weight of user
 
     EnumerableSet.AddressSet private poolInfo;
 
     constructor(
-        address _operatorMsg, address _boost, address __ve, uint _duration
-    )CheckPermission(_operatorMsg) {
+        address _operatorMsg,
+        address _boost,
+        address __ve,
+        uint256 _duration
+    ) CheckPermission(_operatorMsg) {
         veToken = __ve;
         base = IVeToken(__ve).token();
         distribute = _boost;
         duration = _duration;
     }
 
-    function setDuration(uint _duration) external onlyOperator {
+    function setDuration(uint256 _duration) external onlyOperator {
         duration = _duration;
     }
 
-    function reset(uint _tokenId) external {
+    function reset(uint256 _tokenId) external {
         require(IVeToken(veToken).isApprovedOrOwner(msg.sender, _tokenId));
         PoolVote storage poolVote = userPool[_tokenId];
         require(poolVote.lastUse + duration < block.timestamp, "next duration use");
@@ -66,8 +66,8 @@ abstract contract AbstractController is CheckPermission {
         updatePool();
     }
 
-    function _reset(uint _tokenId) internal {
-        uint _totalWeight = usedWeights[_tokenId];
+    function _reset(uint256 _tokenId) internal {
+        uint256 _totalWeight = usedWeights[_tokenId];
         address _pool = userPool[_tokenId].pool;
         emit Abstained(_tokenId, _totalWeight);
         totalWeight -= _totalWeight;
@@ -75,9 +75,9 @@ abstract contract AbstractController is CheckPermission {
         delete userPool[_tokenId];
     }
 
-    function _vote(uint _tokenId, address _poolVote) internal {
+    function _vote(uint256 _tokenId, address _poolVote) internal {
         _reset(_tokenId);
-        uint _weight = IVeToken(veToken).balanceOfNFT(_tokenId);
+        uint256 _weight = IVeToken(veToken).balanceOfNFT(_tokenId);
 
         weights[_poolVote] = weights[_poolVote].add(_weight);
         emit Voted(msg.sender, _tokenId, _weight);
@@ -87,12 +87,11 @@ abstract contract AbstractController is CheckPermission {
         updatePool();
     }
 
-    function poke(uint _tokenId, address _pool) external {
+    function poke(uint256 _tokenId, address _pool) external {
         _vote(_tokenId, _pool);
     }
 
-
-    function vote(uint tokenId, address _poolVote) external {
+    function vote(uint256 tokenId, address _poolVote) external {
         require(IVeToken(veToken).isApprovedOrOwner(msg.sender, tokenId));
         PoolVote storage poolVote = userPool[tokenId];
         require(poolVote.lastUse + duration < block.timestamp, "next duration use");
@@ -112,13 +111,11 @@ abstract contract AbstractController is CheckPermission {
         }
         IDistribute(distribute).massUpdatePools();
         lastUpdate = block.timestamp;
-
     }
 
     function addPool(address _address) external onlyOperator {
         require(_address != address(0), "0 address");
         EnumerableSet.add(poolInfo, _address);
-
     }
 
     function removePool(address _address) external onlyOperator {
@@ -138,9 +135,8 @@ abstract contract AbstractController is CheckPermission {
         return EnumerableSet.contains(poolInfo, _pool);
     }
 
-    function getPool(uint256 _index) public view returns (address){
+    function getPool(uint256 _index) public view returns (address) {
         require(_index <= getPoolLength() - 1, ": index out of bounds");
         return EnumerableSet.at(poolInfo, _index);
     }
-
 }
