+++ contracts/interface/IGauge.sol
@@ -3,11 +3,11 @@
 pragma solidity 0.8.10;
 
 interface IGauge {
-    function notifyRewardAmount(address token, uint amount) external;
+    function notifyRewardAmount(address token, uint256 amount) external;
 
     function getReward(address account) external;
 
-    function claimFees() external returns (uint claimed0, uint claimed1);
+    function claimFees() external returns (uint256 claimed0, uint256 claimed1);
 
-    function left(address token) external view returns (uint);
+    function left(address token) external view returns (uint256);
 }
