+++ contracts/tools/Timelock.sol
@@ -5,34 +5,48 @@ pragma solidity 0.8.4;
 import "@openzeppelin/contracts/utils/math/SafeMath.sol";
 
 contract Timelock {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     event NewAdmin(address indexed newAdmin);
     event NewPendingAdmin(address indexed newPendingAdmin);
-    event NewDelay(uint indexed newDelay);
+    event NewDelay(uint256 indexed newDelay);
     event CancelTransaction(
-        bytes32 indexed txHash, address indexed target, uint value, string signature, bytes data, uint eta
+        bytes32 indexed txHash,
+        address indexed target,
+        uint256 value,
+        string signature,
+        bytes data,
+        uint256 eta
     );
     event ExecuteTransaction(
-        bytes32 indexed txHash, address indexed target, uint value, string signature, bytes data, uint eta
+        bytes32 indexed txHash,
+        address indexed target,
+        uint256 value,
+        string signature,
+        bytes data,
+        uint256 eta
     );
     event QueueTransaction(
-        bytes32 indexed txHash, address indexed target, uint value, string signature, bytes data, uint eta
+        bytes32 indexed txHash,
+        address indexed target,
+        uint256 value,
+        string signature,
+        bytes data,
+        uint256 eta
     );
 
-    uint public constant GRACE_PERIOD = 14 days;
-    uint public constant MINIMUM_DELAY = 12 hours;
-    uint public constant MAXIMUM_DELAY = 30 days;
+    uint256 public constant GRACE_PERIOD = 14 days;
+    uint256 public constant MINIMUM_DELAY = 12 hours;
+    uint256 public constant MAXIMUM_DELAY = 30 days;
 
     address public admin;
     address public pendingAdmin;
-    uint public delay;
+    uint256 public delay;
     bool public admin_initialized;
 
     mapping(bytes32 => bool) public queuedTransactions;
 
-
-    constructor(address admin_, uint delay_) public {
+    constructor(address admin_, uint256 delay_) public {
         require(delay_ <= MAXIMUM_DELAY, "Timelock::constructor: Delay must not exceed maximum delay.");
         admin = admin_;
         delay = delay_;
@@ -42,7 +56,7 @@ contract Timelock {
     // XXX: function() external payable { }
     receive() external payable {}
 
-    function setDelay(uint delay_) public {
+    function setDelay(uint256 delay_) public {
         require(msg.sender == address(this), "Timelock::setDelay: Call must come from Timelock.");
         require(delay_ >= MINIMUM_DELAY, "Timelock::setDelay: Delay must exceed minimum delay.");
         require(delay_ <= MAXIMUM_DELAY, "Timelock::setDelay: Delay must not exceed maximum delay.");
@@ -73,8 +87,12 @@ contract Timelock {
     }
 
     function queueTransaction(
-        address target, uint value, string memory signature, bytes memory data, uint eta
-    ) public returns (bytes32){
+        address target,
+        uint256 value,
+        string memory signature,
+        bytes memory data,
+        uint256 eta
+    ) public returns (bytes32) {
         require(msg.sender == admin, "Timelock::queueTransaction: Call must come from admin.");
         require(
             eta >= getBlockTimestamp().add(delay),
@@ -89,7 +107,11 @@ contract Timelock {
     }
 
     function cancelTransaction(
-        address target, uint value, string memory signature, bytes memory data, uint eta
+        address target,
+        uint256 value,
+        string memory signature,
+        bytes memory data,
+        uint256 eta
     ) public {
         require(msg.sender == admin, "Timelock::cancelTransaction: Call must come from admin.");
 
@@ -100,7 +122,11 @@ contract Timelock {
     }
 
     function executeTransaction(
-        address target, uint value, string memory signature, bytes memory data, uint eta
+        address target,
+        uint256 value,
+        string memory signature,
+        bytes memory data,
+        uint256 eta
     ) public payable returns (bytes memory) {
         require(msg.sender == admin, "Timelock::executeTransaction: Call must come from admin.");
 
@@ -120,7 +146,7 @@ contract Timelock {
         }
 
         // solium-disable-next-line security/no-call-value
-        (bool success, bytes memory returnData) = target.call{value : value}(callData);
+        (bool success, bytes memory returnData) = target.call{value: value}(callData);
         require(success, "Timelock::executeTransaction: Transaction execution reverted.");
 
         emit ExecuteTransaction(txHash, target, value, signature, data, eta);
@@ -128,7 +154,7 @@ contract Timelock {
         return returnData;
     }
 
-    function getBlockTimestamp() internal view returns (uint) {
+    function getBlockTimestamp() internal view returns (uint256) {
         // solium-disable-next-line security/no-block-members
         return block.timestamp;
     }
